+++ fake86/fake86.ino
@@ -1,8 +1,9 @@
-#include "interupt.c"
+#include "cpu8086.h"
+#include "outputserial.h"
 
 void setup() {
   // put your setup code here, to run once:
-  Serial.begin(115200);
+  Serial.begin(9600);
   init86();
 }
