+++ fake86/outputserial.h
@@ -1,0 +1,7 @@
+/*
+    fake86 port on Arduino
+    outputserial.h - From https://stackoverflow.com/questions/66632376/how-to-call-serial-print-from-c-file-in-arduino-ide
+    Copyright (C) 2021 @raspiduino
+*/
+
+void my_log(char *msg);
